Use cats-effect `bracket` over custom `guarantee`
Use `bracket` from the cats-effect Bracket typeclass instead of a custom
`guarantee` method. Bracket also has a `guarantee` method but it is best
practive to use `bracket` "for the acquisition and release of
resources".

As all internal free monads have an Async instance no additional
constraints are required. However `trans` now requires a Bracket instead
of a Monad and `exec` requires a Bracket instead of a MonadError.