Make Query#executeQuery cancelable
The reason of Query#executeQuery to be uncancelable is use of bracket to release ResultSet. But we don't need such strong guarantee, because we can rely on PreparedStatement#close() (https://docs.oracle.com/javase/7/docs/api/java/sql/Statement.html#close()). So, it's possible to use guarantee, which doesn't prevent cancelation.