Fix effectful `copyIn` (Fixes #1512)
The challenge faced here was to capture the number of affected rows when ending a `COPY`.
To safely end a copy, it needs to be run in the finalizer of a `bracket`. We cannot retrieve values from a finalizer, so there was some logic in place to call `PFCI.endCopy` in two places and use `isActive` to determine if it should be called the second time. `isActive` seems to not be an `!isClosed`, but rather "is connection reserved for this Copy operation?".

The fix is to replace the double closing with a reference